# apps/web/src/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "ClosureRulesResponse": {
        "properties": {
          "close_on_contract_status": {
            "title": "Close On Contract Status",
            "type": "boolean"
          },
          "close_on_winner_status": {
            "title": "Close On Winner Status",
            "type": "boolean"
          },
          "consulting_timeout_days": {
            "title": "Consulting Timeout Days",
            "type": "integer"
          },
          "contract_status_terms": {
            "items": {
              "type": "string"
            },
            "title": "Contract Status Terms",
            "type": "array"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "stale_eligible_states": {
            "items": {
              "type": "string"
            },
            "title": "Stale Eligible States",
            "type": "array"
          },
          "stale_no_tor_days": {
            "title": "Stale No Tor Days",
            "type": "integer"
          },
          "winner_status_terms": {
            "items": {
              "type": "string"
            },
            "title": "Winner Status Terms",
            "type": "array"
          }
        },
        "required": [
          "close_on_winner_status",
          "close_on_contract_status",
          "winner_status_terms",
          "contract_status_terms",
          "consulting_timeout_days",
          "stale_no_tor_days",
          "stale_eligible_states",
          "source"
        ],
        "title": "ClosureRulesResponse",
        "type": "object"
      },
...
      "CreateRuleProfileRequest": {
        "properties": {
          "close_consulting_after_days": {
            "default": 30,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Close Consulting After Days",
            "type": "integer"
          },
          "close_stale_after_days": {
            "default": 45,
            "maximum": 365.0,
            "minimum": 1.0,
            "title": "Close Stale After Days",
            "type": "integer"
          },
          "enabled_by_user": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled By User"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "max_pages_per_keyword": {
            "default": 15,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Pages Per Keyword",
            "type": "integer"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "profile_type": {
            "default": "custom",
            "minLength": 1,
            "title": "Profile Type",
            "type": "string"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateRuleProfileRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManualRecrawlQueuedResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "inflight_run_count": {
            "title": "Inflight Run Count",
            "type": "integer"
          },
          "max_concurrent_runs": {
            "title": "Max Concurrent Runs",
            "type": "integer"
          },
          "max_queued_keywords": {
            "title": "Max Queued Keywords",
            "type": "integer"
          },
          "queued_keyword_count": {
            "title": "Queued Keyword Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "code",
          "status",
          "inflight_run_count",
          "max_concurrent_runs",
          "queued_keyword_count",
          "max_queued_keywords"
        ],
        "title": "ManualRecrawlQueuedResponse",
        "type": "object"
      },
      "ManualRecrawlRequest": {
        "properties": {
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "title": "ManualRecrawlRequest",
        "type": "object"
      },
      "ManualRecrawlResponse": {
        "properties": {
          "queued_job_count": {
            "title": "Queued Job Count",
            "type": "integer"
          },
          "queued_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Queued Keywords",
            "type": "array"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "queued_job_count",
          "queued_keywords"
        ],
        "title": "ManualRecrawlResponse",
        "type": "object"
      },
      "ManualRecrawlStatusResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "failed_count": {
            "title": "Failed Count",
            "type": "integer"
          },
          "failed_keywords": {
            "items": {
              "type": "string"
            },
            "title": "Failed Keywords",
            "type": "array"
          },
          "is_terminal": {
            "title": "Is Terminal",
            "type": "boolean"
          },
          "partial_count": {
            "title": "Partial Count",
            "type": "integer"
          },
          "queued_count": {
            "title": "Queued Count",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "requested_keyword_count": {
            "title": "Requested Keyword Count",
            "type": "integer"
          },
          "retrying_count": {
            "title": "Retrying Count",
            "type": "integer"
          },
          "running_count": {
            "title": "Running Count",
            "type": "integer"
          },
          "succeeded_count": {
            "title": "Succeeded Count",
            "type": "integer"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "zero_result_count": {
            "title": "Zero Result Count",
            "type": "integer"
          }
        },
        "required": [
          "request_id",
          "requested_keyword_count",
          "queued_count",
          "running_count",
          "retrying_count",
          "succeeded_count",
          "zero_result_count",
          "partial_count",
          "failed_count",
          "failed_keywords",
          "is_terminal",
          "created_at",
          "updated_at"
        ],
        "title": "ManualRecrawlStatusResponse",
        "type": "object"
      },
...
      "NotificationRulesResponse": {
        "properties": {
          "event_wiring_complete": {
            "title": "Event Wiring Complete",
            "type": "boolean"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "supported_channels": {
            "items": {
              "type": "string"
            },
            "title": "Supported Channels",
            "type": "array"
          },
          "supported_types": {
            "items": {
              "type": "string"
            },
            "title": "Supported Types",
            "type": "array"
          }
        },
        "required": [
          "supported_channels",
          "supported_types",
          "event_wiring_complete",
          "source"
        ],
        "title": "NotificationRulesResponse",
        "type": "object"
      },
...
      "RuleProfileResponse": {
        "properties": {
          "close_consulting_after_days": {
            "title": "Close Consulting After Days",
            "type": "integer"
          },
          "close_stale_after_days": {
            "title": "Close Stale After Days",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "effective_status": {
            "title": "Effective Status",
            "type": "string"
          },
          "enabled_by_user": {
            "title": "Enabled By User",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "title": "Keywords",
            "type": "array"
          },
          "max_pages_per_keyword": {
            "title": "Max Pages Per Keyword",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "profile_type": {
            "title": "Profile Type",
            "type": "string"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "profile_type",
          "enabled_by_user",
          "is_active",
          "effective_status",
          "status_reason",
          "max_pages_per_keyword",
          "close_consulting_after_days",
          "close_stale_after_days",
          "keywords",
          "created_at",
          "updated_at"
        ],
        "title": "RuleProfileResponse",
        "type": "object"
      },
      "RulesResponse": {
        "properties": {
          "closure_rules": {
            "$ref": "#/components/schemas/ClosureRulesResponse"
          },
          "entitlements": {
            "$ref": "#/components/schemas/EntitlementSummaryResponse"
          },
          "notification_rules": {
            "$ref": "#/components/schemas/NotificationRulesResponse"
          },
          "profiles": {
            "items": {
              "$ref": "#/components/schemas/RuleProfileResponse"
            },
            "title": "Profiles",
            "type": "array"
          },
          "schedule_rules": {
            "$ref": "#/components/schemas/ScheduleRulesResponse"
          }
        },
        "required": [
          "profiles",
          "entitlements",
          "closure_rules",
          "notification_rules",
          "schedule_rules"
        ],
        "title": "RulesResponse",
        "type": "object"
      },
...
      "UpdateRuleProfileRequest": {
        "properties": {
          "close_consulting_after_days": {
            "anyOf": [
              {
                "maximum": 365.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Consulting After Days"
          },
          "close_stale_after_days": {
            "anyOf": [
              {
                "maximum": 365.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Stale After Days"
          },
          "enabled_by_user": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled By User"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "max_pages_per_keyword": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pages Per Keyword"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          }
        },
        "title": "UpdateRuleProfileRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Thailand public procurement monitoring API",
    "title": "e-GP Intelligence Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/rules": {
      "get": {
        "operationId": "get_rules_v1_rules_get",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Rules",
        "tags": [
          "rules"
        ]
      }
    },
    "/v1/rules/profiles": {
      "post": {
        "operationId": "create_rule_profile_v1_rules_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRuleProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule Profile",
        "tags": [
          "rules"
        ]
      }
    },
    "/v1/rules/profiles/{profile_id}": {
      "patch": {
        "operationId": "update_rule_profile_v1_rules_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRuleProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule Profile",
        "tags": [
          "rules"
        ]
      }
    },
    "/v1/rules/recrawl": {
      "post": {
        "operationId": "recrawl_active_keywords_v1_rules_recrawl_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualRecrawlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualRecrawlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualRecrawlQueuedResponse"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Recrawl Active Keywords",
        "tags": [
          "rules"
        ]
      }
    },
    "/v1/rules/recrawl/{request_id}": {
      "get": {
        "operationId": "get_recrawl_request_status_v1_rules_recrawl__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManualRecrawlStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recrawl Request Status",
        "tags": [
          "rules"
        ]
      }
    },
...
  }
}